requirements.txt
asgiref
Django
djangorestframework
pytz
sqlparse
psycopg2-binary
djangorestframework-simplejwt
drf-yasg
django-cors-headers
django-cors-middleware
django-mptt
pytest
pytest-django
pytest-cov
mixer
mock